updated with pylint
