+++ update_users.py
@@ -1,9 +1,11 @@
-import pymongo
+"""
+This class handles the updating of users with the database
+"""
 from pymongo import MongoClient
 
 # Connecting to database and declaring the collection
-cluster = MongoClient("mongodb+srv://buzzbot:neJPIzxDGax66NqI@buzzbot-" 
-+ "gesgl.gcp.mongodb.net/test?retryWrites=true&w=majority")
+cluster = MongoClient("mongodb+srv://buzzbot:neJPIzxDGax66NqI@buzzbot-"
+                      + "gesgl.gcp.mongodb.net/test?retryWrites=true&w=majority")
 db = cluster["discord"]
 collection = db["users"]
 
@@ -14,46 +16,46 @@ async def first_time(self):
     :param self: this is our bot
     """
     # x is a list created by the method get_all_members()
-    x = self.get_all_members()
-    for member in x:
+    member_list = self.get_all_members()
+    for member in member_list:
         post = {
-        "_id": member.id, 
-        "name": member.name,
-        "joined at": member.joined_at,
-        "server": member.guild.name, 
-        "isBot": member.bot
+            "_id": member.id,
+            "name": member.name,
+            "joined at": member.joined_at,
+            "server": member.guild.name,
+            "isBot": member.bot
         }
         collection.insert_one(post)
 
+
 async def new_member(member):
     """
-    Adds newly joined memebers while bot is on to the database
-    :param member: this is the memeber that has just joined
+    Adds newly joined members while bot is on to the database
+    :param member: this is the member that has just joined
     """
     post = {
-        "_id": member.id, 
+        "_id": member.id,
         "name": member.name,
         "joined at": member.joined_at,
-        "server": member.guild.name, 
+        "server": member.guild.name,
         "isBot": member.bot
-        }
+    }
     collection.insert_one(post)
+
 
 async def remove_member(member):
     """
     Removes member from database when they leave the server
     :param member: member that left
     """
-    leaving_member = { "_id": member.id}
+    leaving_member = {"_id": member.id}
     collection.delete_one(leaving_member)
+
 
 async def remove_channel(guild):
     """
     Removes member from database when they leave the server
-    :param member: member that left
+    :param guild: server that removed the bot
     """
-    leaving_server = { "server": guild.name}
+    leaving_server = {"server": guild.name}
     collection.delete_one(leaving_server)
-
-
-
